Тесты: изменен тип данных с float на double

+++ FHE/UnitTest/UnitTest1.cs
@@ -108,12 +108,12 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc1()
         {
-            float correctlyResult = 4;
+            double correctlyResult = 4;
             Function func = new Function("1 + x1");
-            Dictionary<String, float> args = new Dictionary<string,float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 3);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func 1+x1 not correct");
         }
@@ -122,13 +122,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc2()
         {
-            float correctlyResult = 13.5f;
+            double correctlyResult = 13.5f;
             Function func = new Function("x1+x2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 4.5f);
             args.Add("x2", 9);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1+x2 not correct");
         }
@@ -137,12 +137,12 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc3()
         {
-            float correctlyResult = 0.1f;
+            double correctlyResult = 0.1f;
             Function func = new Function("1-x1");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 0.9f);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func 1-x1 not correct");
         }
@@ -151,13 +151,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc4()
         {
-            float correctlyResult = 26;
+            double correctlyResult = 26;
             Function func = new Function("x1-x2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 56);
             args.Add("x2", 30);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1-x2 not correct");
         }
@@ -166,12 +166,12 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc5()
         {
-            float correctlyResult = 27;
+            double correctlyResult = 27;
             Function func = new Function("3*x1");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 9);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func 3*x1 not correct");
         }
@@ -180,13 +180,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc6()
         {
-            float correctlyResult = 0.9f;
+            double correctlyResult = 0.9f;
             Function func = new Function("x1*x2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 0.1f);
             args.Add("x2", 9);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1*x2 not correct");
         }
@@ -195,12 +195,12 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc7()
         {
-            float correctlyResult = 4.5f;
+            double correctlyResult = 4.5f;
             Function func = new Function("x1/2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 9);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1/2 not correct");
         }
@@ -209,13 +209,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc8()
         {
-            float correctlyResult = 1;
+            double correctlyResult = 1;
             Function func = new Function("x1/x2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 3);
             args.Add("x2", 3);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1/x2 not correct");
         }
@@ -224,12 +224,12 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc9()
         {
-            float correctlyResult = 1;
+            double correctlyResult = 1;
             Function func = new Function("x1^2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 1);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1^2 not correct");
         }
@@ -238,13 +238,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc10()
         {
-            float correctlyResult = 3;
+            double correctlyResult = 3;
             Function func = new Function("x1^x2");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 9);
             args.Add("x2", 0.5f);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func x1^x2 not correct");
         }
@@ -253,13 +253,13 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc11()
         {
-            float correctlyResult = 8;
+            double correctlyResult = 8;
             Function func = new Function("3*x1+x2/4");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 2);
             args.Add("x2", 8);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func 3*x1+x2/4 not correct");
         }
@@ -268,14 +268,14 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc12()
         {
-            float correctlyResult = 15.466f;
+            double correctlyResult = 15.466f;
             Function func = new Function("(x1+x2+x3)/3");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 3);
             args.Add("x2", 9.4f);
             args.Add("x3", 34);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func (x1+x2+x3)/3 not correct");
         }
@@ -284,14 +284,14 @@ namespace UnitTest
         [TestMethod]
         public void testCalcResultFunc13()
         {
-            float correctlyResult = 16807;
+            double correctlyResult = 16807;
             Function func = new Function("(3*x1-x2)^5*x3");
-            Dictionary<String, float> args = new Dictionary<string, float>();
+            Dictionary<String, double> args = new Dictionary<string, double>();
             args.Add("x1", 4);
             args.Add("x2", 5);
             args.Add("x3", 1);
 
-            float result = func.calcResult(args);
+            double result = func.calcResult(args);
 
             Assert.AreEqual(result, correctlyResult, 0.001, "Func (3*x1-x2)^5*x3 not correct");
         }
